Created service object for csv
