+++ app/models/map_asset.rb
@@ -3,7 +3,7 @@ class MapAsset < ApplicationRecord
   include PgSearch::Model
   include Scopes
   include Validations
-  include Functions
+  include Attributes
 
   # Associations
   has_rich_text :description
@@ -11,5 +11,9 @@ class MapAsset < ApplicationRecord
 
   belongs_to :user
   has_many :stories, dependent: :destroy
+
+  def self.send_chain(methods)
+    methods.inject(self) { |result, method| result.send(*method) }
+  end
 
 end
